+++ packages/partials/ts-types/src/Typings.ts
@@ -141,3 +141,16 @@ export type IfIsNever<T, TYes, TNo> = [T] extends [never] ? TYes : TNo;
  * @see {@link IConstructor}
  */
 export type IInstanceOf<T> = T extends IConstructor<infer U> ? U : never;
+
+export type IDeepReadonly<T extends IObject | any[]> = {
+
+    /* eslint-disable @stylistic/indent */
+
+    readonly [P in keyof T]: T[P] extends IObject ?
+        IDeepReadonly<T[P]>
+        : T[P] extends Array<infer U> ?
+            U extends IObject ? ReadonlyArray<IDeepReadonly<U>> : readonly U[]
+            : T[P];
+
+    /* eslint-enable @stylistic/indent */
+};